updated split/combine cubes logic
based on the idea of extended cube, a brand new method for splitting and combining parts of a cube was implemented. Now from the basic cube (ex. 6x6x6) we extract the internal pieces (x8 - 2x2x2) as a variable parts and other external pieces as a constants. We send to the nodes cubes that have both - external and internal part (x8 - 4x4x4 -node's cube). Each nodes (processes) updates the values of internal part node's cube then extract internal part as a separate cube and send it to the master.
Master combine internal cubes into one.
TO-DO: next step is to update internal part of the basic cube based on the recived internal cube.